+++ __tests__/playwright/work-experience/projectItemLinksWork.spec.ts
@@ -1,6 +1,6 @@
 import { testProjectLink } from '@/__tests__/playwright/utils/testProjectLink'
+import { PROJECT_ID } from '@/utils/constants/ids/projectIds'
 import { PAGES_URL } from '@/utils/constants/pageUrls'
-import { PROJECT_ID } from '@/utils/constants/projectIds'
 import { Browser, BrowserContext, Page, chromium, test } from '@playwright/test'
 
 let browser: Browser
