+++ app/error.tsx
@@ -1,7 +1,7 @@
 'use client' // Error components must be Client Components
 
 import { ErrorPageLayout } from '@/components/layout/ErrorPageLayout'
-import { ID } from '@/utils/constants/ids'
+import { ID } from '@/utils/constants/ids/ids'
 import { ErrorProps } from '@/utils/shared/componentProps'
 
 export default function Error({ error, reset }: ErrorProps) {
