+++ app/work-experience/page.tsx
@@ -13,7 +13,7 @@ import {
   projectsWorkWordPress,
 } from '@/data/projects/work/projectsOverview'
 import { TEXT } from '@/localization/texts_en'
-import { ID } from '@/utils/constants/ids'
+import { ID } from '@/utils/constants/ids/ids'
 import { PAGES_URL } from '@/utils/constants/pageUrls'
 import { DATA_TEST_IDS } from '@/utils/dataTestIds'
 import { Metadata } from 'next'
