+++ app/work-experience/smartsupp-help/page.tsx
@@ -3,7 +3,7 @@ import PageNavigation from '@/components/pageNavigation/PageNavigation'
 import { projectsWorkWordPress } from '@/data/projects/work/projectsOverview/workWordpress'
 import { sections } from '@/data/projects/work/smartsuppHelp'
 import { TEXT } from '@/localization/texts_en'
-import { ID } from '@/utils/constants/ids'
+import { ID } from '@/utils/constants/ids/ids'
 import { PAGES_URL, URL_WORK_EXPERIENCE } from '@/utils/constants/pageUrls'
 import { DATA_TEST_IDS } from '@/utils/dataTestIds'
 import { BreadCrumbsType } from '@/utils/types'
