+++ data/projects/personal/projectsOverview/personalNext.ts
@@ -9,7 +9,8 @@ import krsiak4 from '@/public/images/projects-personal/krsiak/krsiak-4-light.png
 import krsiak5 from '@/public/images/projects-personal/krsiak/krsiak-5-light.png'
 import krsiakThumb from '@/public/images/projects-personal/krsiak/krsiak-thumb-light-600.webp'
 
-import { PROJECT_ID } from '@/utils/constants/projectIds'
+import { PROJECT_ID } from '@/utils/constants/ids/projectIds'
+import { PAGES_URL } from '@/utils/constants/pageUrls'
 
 export const projectsPersonalNext: Project[] = [
   {
