+++ data/projects/personal/projectsOverview/personalReact.ts
@@ -10,7 +10,8 @@ import cryptomania6 from '@/public/images/projects-personal/cryptomania/cryptoma
 import cryptomania3 from '@/public/images/projects-personal/cryptomania/cryptomania-3-light.png'
 import cryptomaniaThumb from '@/public/images/projects-personal/cryptomania/cryptomania-thumb-light-600.webp'
 
-import { PROJECT_ID } from '@/utils/constants/projectIds'
+import { PROJECT_ID } from '@/utils/constants/ids/projectIds'
+import { PAGES_URL } from '@/utils/constants/pageUrls'
 
 export const projectsPersonalReact: Project[] = [
   {
