+++ data/projects/work/projectsOverview/workFrontEnd.ts
@@ -1,5 +1,5 @@
 import { iconsSkillsProjectsWorkSmartsuppWeb } from '@/data/skills/skillsProjectsWork'
-import { PROJECT_ID } from '@/utils/constants/projectIds'
+import { PROJECT_ID } from '@/utils/constants/ids/projectIds'
 import { Project } from '@/utils/interfaces'
 
 // smartsupp web
@@ -8,6 +8,7 @@ import smartsuppWeb2 from '@/public/images/projects-work/smartsupp-web/smartsupp
 import smartsuppWeb3 from '@/public/images/projects-work/smartsupp-web/smartsupp-web-3-light.png'
 import smartsuppWeb4 from '@/public/images/projects-work/smartsupp-web/smartsupp-web-4-light.png'
 import smartsuppWebThumb from '@/public/images/projects-work/smartsupp-web/smartsupp-web-thumb-light-600.webp'
+import { PAGES_URL } from '@/utils/constants/pageUrls'
 
 export const projectsWorkFrontEnd: Project[] = [
   {
